+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 - .env.local.php in Symfony recipe. [#2506]
 - Fixed regex identifying "cd" commands in YAML scripts. The regex was not taking into account that multiple commands can be executed within the same line — e.g. "cd {{release_path}} && npm run prod". [#2509]
 - Slack default channel null breaks webhook. [#2525]
+- Use port for ssh-keyscan when applicable. [#2549]
 - Support passing `null` as default in get(). [#2545]
 - Shopware recipe: First activate plugins THEN build them to avoid breaking theme:compile with unbuild themes (which were activated AFTER build).
 - Shopware recipe sw:plugin:upgrade:all task.
@@ -625,6 +626,7 @@
 - Fixed `DotArray` syntax in `Collection`.
 
 
+[#2549]: https://github.com/deployphp/deployer/issues/2549
 [#2545]: https://github.com/deployphp/deployer/issues/2545
 [#2525]: https://github.com/deployphp/deployer/issues/2525
 [#2509]: https://github.com/deployphp/deployer/issues/2509